+++ Src/MoneyFox.Ui/InversionOfControl/MoneyFoxConfig.cs
@@ -26,6 +26,7 @@ using Views.Payments.PaymentModification;
 using Views.Settings;
 using Views.Setup;
 using Views.Setup.SelectCurrency;
+using Views.Setup.SetupAccounts;
 using Views.Statistics;
 using Views.Statistics.CashFlow;
 using Views.Statistics.CategoryProgression;
