+++ Src/MoneyFox.Ui/Views/OverflowMenu/OverflowMenuView.xaml
@@ -24,15 +24,15 @@
 
                             <HorizontalStackLayout Spacing="6" Padding="12"
                                                    x:DataType="overflowMenu:OverflowItemViewModel">
-                                <Image HeightRequest="15"
-                                       WidthRequest="15">
+                                <Image HeightRequest="15" WidthRequest="15">
                                     <Image.Source>
                                         <FontImageSource FontFamily="MaterialIcons"
                                                          Glyph="{Binding IconGlyph}"
                                                          Color="{AppThemeBinding Dark={StaticResource White}, Light={StaticResource Gray950}}" />
                                     </Image.Source>
                                 </Image>
-                                <Label Text="{Binding Name}" Style="{DynamicResource TextSubHead}"
+                                <Label Text="{Binding Name}" 
+                                       Style="{DynamicResource TextSubHead}"
                                        VerticalOptions="Center" />
                             </HorizontalStackLayout>
                         </controls:Card>
