+++ Src/MoneyFox.Ui/Views/Setup/SetupAccounts/SetupAccountPage.xaml
@@ -2,10 +2,10 @@
 
 <ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             x:Class="MoneyFox.Ui.Views.Setup.SetupAccountPage"
+             x:Class="MoneyFox.Ui.Views.Setup.SetupAccounts.SetupAccountPage"
              xmlns:resources="clr-namespace:MoneyFox.Ui.Resources.Strings"
-             xmlns:setup="clr-namespace:MoneyFox.Ui.Views.Setup"
-             x:DataType="setup:SetupAccountsViewModel">
+             xmlns:setupAccounts="clr-namespace:MoneyFox.Ui.Views.Setup.SetupAccounts"
+             x:DataType="setupAccounts:SetupAccountsViewModel">
 
     <Grid Padding="15"
           IsClippedToBounds="False"
